// src/api/generated/models/ActivityResponse.ts
export interface ActivityResponse {
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    activityId?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    activityType?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    activitySubject?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    activityDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    activityStatus?: string;
    /**
     * 
     * @type {Date}
     * @memberof ActivityResponse
     */
    activityDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ActivityResponse
     */
    dueDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ActivityResponse
     */
    completedDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof ActivityResponse
     */
    durationMinutes?: number;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    priority?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    relatedType?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    relatedId?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    assignedUserId?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    assignedUserName?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof ActivityResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ActivityResponse
     */
    updatedDate?: Date;
}